Move XMLUtils to "utils" package Add funtion description
