+++ OWL/app/src/main/java/be/ulb/owl/utils/XMLUtils.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package be.ulb.owl.xml;
+package be.ulb.owl.utils;
 
 import android.content.res.AssetManager;
 import android.util.Log;
@@ -53,13 +53,26 @@ public class XMLUtils {
 
         return parser;
     }
-    
-    
+
+    /**
+     * Check if the current XMLPullParser (iterator) is a space or not
+     *
+     * @param parser the parser which must be tested
+     * @return True if it is a space
+     * @throws XmlPullParserException
+     */
     public static boolean isSpace(XmlPullParser parser) throws XmlPullParserException {
         return parser.getEventType() == XmlPullParser.TEXT && 
                 (parser.getName() == null || parser.getName().equalsIgnoreCase(""));
     }
-    
+
+    /**
+     * Remove all space from the current iterator to the next data
+     *
+     * @param parser the parser which must be tested
+     * @throws XmlPullParserException
+     * @throws IOException
+     */
     public static void removeSpace(XmlPullParser parser) 
             throws XmlPullParserException, IOException {
         
@@ -67,7 +80,15 @@ public class XMLUtils {
             parser.next();
         }
     }
-    
+
+    /**
+     * Get the next real element of the parser.  Next element of the XMLPullParser and remove all
+     * space after this next element
+     *
+     * @param parser the parser which must be tested
+     * @throws XmlPullParserException
+     * @throws IOException
+     */
     public static void nextAndRemoveSpace(XmlPullParser parser) 
             throws XmlPullParserException, IOException {
